chore: drawer in create and host pages
